+++ src/Freee.Accounting/Models/PaymentParams.cs
@@ -82,7 +82,7 @@ namespace Freee.Accounting.Models
         /// <param name="date">支払日 (required).</param>
         /// <param name="fromWalletableId">口座ID（from_walletable_typeがprivate_account_itemの場合は勘定科目ID）：payments指定時は必須 (required).</param>
         /// <param name="fromWalletableType">口座区分 (銀行口座: bank_account, クレジットカード: credit_card, 現金: wallet, プライベート資金（法人の場合は役員借入金もしくは役員借入金、個人の場合は事業主貸もしくは事業主借）: private_account_item)：payments指定時は必須 (required).</param>
-        public PaymentParams(int amount = default(int), int companyId = default(int), string date = default(string), int fromWalletableId = default(int), FromWalletableTypeEnum fromWalletableType = default(FromWalletableTypeEnum))
+        public PaymentParams(long amount = default(long), int companyId = default(int), string date = default(string), int fromWalletableId = default(int), FromWalletableTypeEnum fromWalletableType = default(FromWalletableTypeEnum))
         {
             this.Amount = amount;
             this.CompanyId = companyId;
@@ -97,7 +97,7 @@ namespace Freee.Accounting.Models
         /// </summary>
         /// <value>金額</value>
         [DataMember(Name = "amount", IsRequired = true, EmitDefaultValue = false)]
-        public int Amount { get; set; }
+        public long Amount { get; set; }
 
         /// <summary>
         /// 事業所ID
